+++ models/user-model.ts
@@ -3,14 +3,14 @@ import { Schema, model } from "mongoose";
 import { IUser } from "../interfaces/user-interface";
 
 const userSchema = new Schema<IUser>({
-  avatar: { type: String, requiered: true },
-  confirmPassword: { type: String, required: true },
-  apellido: { type: String, required: true },
   nombre: { type: String, required: true },
-  telefono_personal: { type: Number, required: true },
-  password: { type: String, required: true },
-  rol: { type: Number, required: true },
+  apellido: { type: String, required: true },
   email: { type: String, required: true, unique: true },
+  password: { type: String, required: true },
+  confirmPassword: { type: String, required: true },
+  telefono_personal: { type: Number, required: true },
+  rol: { type: Number, required: true },
+  avatar: { type: String, requiered: false },
 });
 const User = model<IUser>("User", userSchema);
 export const models = { User };
