+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@hapi/boom": "^10.0.0",
     "bcryptjs": "^2.4.3",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "joi": "^17.6.3",
