+++ routes/usersRoute.ts
@@ -1,20 +1,14 @@
 import express from "express";
-import { createUser, loginUser } from "../controllers/users/users-controller";
-import { userSchema, loginUserSchema } from "../schemas/userSchema";
+import { createUser } from "../controllers/users/users-controller";
 import { validationHandler } from "../middlewares/validationHandler";
+import { userSchema } from "../schemas/userSchema";
 
 const routerUser = express.Router();
 
 routerUser.post(
-  "/create-user",
+  "/",
   validationHandler(userSchema, "body"),
   createUser
-);
-
-routerUser.post(
-  "/AuthReq",
-  validationHandler(loginUserSchema, "body"),
-  loginUser
 );
 
 export default routerUser;
